% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binary__Gibbs_TraceplotsPlots.R
\name{BinaryGibbs_Traceplot}
\alias{BinaryGibbs_Traceplot}
\title{Plots Posterior Frequency Distribution .}
\usage{
BinaryGibbs_Traceplot(beta_matrix, k)
}
\arguments{
\item{beta_matrix}{a nIter X (p+1) matrix of beta_updates.}

\item{k}{a integer not greater than (p+1) indicating which beta is of interest.}
}
\value{
\code{traceplot} Line diagrams showing convergence of gibbs sampler for a parameter and indicating cumulative posterior mean over iterartions.
}
\description{
\code{BinaryGibbs_PosteriorDistribution_plot} Plots Posterior Frequency Distribution of Parameters estimated by Probit Regression for Binary Responses via data augmentation and Gibbs sampling.
}
\examples{
set.seed(250)
require(truncnorm)
require(MASS)
N <- 500
x1 <- seq(-1, 1, length.out = N)
x2 <- rnorm(N, 0, 1)
D <- 3
X <- matrix(c(rep(1, N), x1, x2), ncol = D)
true_theta <- c(- 0.5, 3.3, 2)
p <- pnorm(X \%*\% true_theta)
y <- rbinom(N, 1, p)
N1  <- sum(y)  # Number of successes
N0  <- N - N1  # Number of failures
#Spliting The Data in Train and Test in 80:20 ratio
Train_ID = sample(1:nrow(X), round(nrow(X) * 0.8), replace = FALSE) # Train Data IDS
Train_X = X[Train_ID, -1] # Train Data Covariates
Test_X = X[-Train_ID, -1] # Test Data Covarites
Train_Y = y[Train_ID] # Train Data Response
Test_Y = y[-Train_ID] # Test Data Response
nIter = 10000
burn_in = round(nIter * 0.5)
prior = 2
prior_mean = rep(1, 3)
prior_var = diag(10, 3)
temp = BinaryGibbs_fit(Train_X, Train_Y, nIter, prior, burn_in, prior_mean, prior_var )
BinaryGibbs_Traceplot(beta_matrix = temp$beta_matrix, k = 0)

}
