% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poly_Gibbs_GammaPosterior.R
\name{Multinom_PosteriorDistribution_plot_gamma}
\alias{Multinom_PosteriorDistribution_plot_gamma}
\title{Plots Posterior Frequency Distribution}
\usage{
Multinom_PosteriorDistribution_plot_gamma(gamma_update, k, burn_in, breaks)
}
\arguments{
\item{gamma_update}{a nIter X (K+1) matrix of beta_updates staring with -Inf and ending with + Inf, sorted in each row}

\item{k}{a integer not greater than (p+1) indicating which beta is of interest}

\item{burn_in}{burn_in period , less than (nrow(beta_matrix) - 1)}

\item{breaks}{integer, no of breaks in histogram}
}
\value{
\code{PosteriorDistribution_plot} A histrogram showing Posterior Frequency Distribution and Posterior Mean
}
\description{
\code{Multinom_PosteriorDistribution_plot_beta} Plots Posterior Frequency Distribution of Boundary Parameters estimated by Ordered Multinomial Regression via data augmentation and Gibbs sampling.
}
\examples{
# Initialization
set.seed(250)
n <- 1000 # Total no of observations.
int1 <- -1 # gamma boundary
int2 <- 3  # gamma boundary
beta <- c(-.75, 1) # Regression Parameters for data generation.
X <- cbind(sample(1:4, n, replace = TRUE), rnorm(n, 0, 2)) # Generated design matrix
# Generation of Latent Variable Observations
eta <- X \%*\% beta
z <- rnorm(n, eta, 1)
# Generation of Responses depending on z
y <- rep(0, n)
y[z <= int1] <- 1
y[int1 <z & z <= int2] <- 2
y[int2 < z ] <- 3
#Spliting The Data in Train and Test in 80:20 ratio
Train_ID = sample(1:nrow(X), round(nrow(X) * 0.8), replace = FALSE) # Train Data IDS
Train_X = X[Train_ID, ]# Train Data Covariates
Test_X = X[-Train_ID, ]
Train_Y = y[Train_ID] # Train Data Response
Test_Y = y[-Train_ID] # Test Data Response
K = 3
nIter = 10000
burn_in = 5000
breaks = 50
Result = MultinomGibbs_fit(Train_X, Train_Y, nIter, burn_in, K)
gamma_update = Result$gamma_update
Multinom_PosteriorDistribution_plot_gamma(gamma_update = gamma_update , k = 2, burn_in = 2500, breaks= 50)
}
