% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Split.R
\name{split}
\alias{split}
\title{Spliting Data in Trainset and Test Set.}
\usage{
split(X, y, Trainset_Proportion)
}
\arguments{
\item{X}{A N X p matrix of covariates}

\item{y}{A N X 1 vector of responses}

\item{Trainset_Percentage:}{(A real no between 0 and 1 both exclusive)}
}
\value{
\code{Train_X} A [N * Trainset_Proportion] X p matrix of covariates.

\code{Train_Y} A [N * Trainset_Proportion] X 1 vector of responses.

\code{Test_X} A [N * (1 - Trainset_Proportion)] X p matrix of covariates.

\code{Test_Y} A [N * (1 - Trainset_Proportion)] X 1 vector of responses.
}
\description{
\code{Split} Splits Data in Trainset and Test Set
}
\examples{
set.seed(250)
N <- 500 # Total no of observations.
x1 <- seq(-1, 1, length.out = N) # Generating Covariate 1
x2 <- rnorm(N, 0, 1) # Genearting Covariate 2
# Generating the design matrix of interest
X <- as.matrix(cbind(rep(1,N), x1, x2))
# Fixing values of regression coeffiecients theta
theta <- c(-.5, 3.3, 2)
# Define a vector with probabilities of success p using the probit link
p <- pnorm(X \%*\% theta)
# Generate binary response data y with success probabilit p
y <- rbinom(N, 1, p)
split(X, y, Trainset_Proportion = .7)

}
